refactor: onboarding functionality to new controller
Previously, onboarding was taken care by profile controller. Since onboarding is a one time task, it will not be efficient to use it in profile controller which gets created every time the user uses the app or goes to profile screen.
