--- lib/controllers/example_controller.dart
@@ -1,8 +1,0 @@
-import 'package:get/get.dart';
-
-//TODO: Delete this file after adding any file to this directory
-
-class ExampleController extends GetxController
-{
-  var example = 1.obs;
-}
